reverse bar log entries through a separate page instead of a form
The bar log contained a form with one submit button for each log entry.
This was used as a button to revert the log entry, using the buttons value
to store the id of the log entry to be reverted.

This commit changes the revert button to be a link to a new page,
/bar/reverse_<bar entry id>, which will delete the entry and redirect to
the bar log. If the argument 'prev' is sent along with the request
the page will redirect to this url. This way we can ensure the user is
sent back to the exact page she was on when clicking the revert button,
including pagination and filters.