Save the start and end dates of membership in the database.
We now save a membership's start and end date instead of whether a user is
a member. This allows us to distinguish between former members and new
member applications. User.active_member is still accessible but computed
based on the saved dates. Saving the dates will also allow us to compute
who our members were at a previous point in time, which is a legal
requirement.